Refactor: Improve code structure, readability, and remove redundancies
This commit introduces significant refactoring to the codebase with the primary goals of improving readability, reducing code duplication, and making the component structure more modular and maintainable.

Key changes include:

- **Linting:** Addressed all linting warnings, including removal of unused variables.
- **Component Extraction:**
    - The main `App.jsx` component was broken down by extracting several UI sections and page views into their own dedicated components within `src/components/`:
        - `LatestResults.jsx`: Displays the latest game results on the home page.
        - `SchedulePage.jsx`: Displays the upcoming schedule.
        - `MatchdayResultsView.jsx`: Handles display and navigation of results per matchday.
        - `TeamResultsView.jsx`: Handles display of results filtered by team.
    - The `ResultsByMatchday` component in `App.jsx` now serves as a container for `MatchdayResultsView` and `TeamResultsView`.
- **Utility Functions:**
    - Created `src/utils.js` to house common helper functions:
        - `getUniqueGames`: Centralized logic to deduplicate game entries.
        - `calculateStandings`: Moved from `App.jsx`, calculates league standings.
        - `getDynamicHighlights`: Moved from `App.jsx`, generates highlight strings.
- **State Management Simplification:**
    - Removed redundant `teams` state from `App.jsx` as the data was used locally within the data loading effect.
- **Code Cleanup:**
    - Removed unused commented-out code.
    - Ensured consistency in `db.js` (though no changes were needed here).

These changes make the codebase easier to understand, navigate, and maintain without altering existing application functionality.
